Add files via upload
A method that takes two Strings as input and returns the String whose 
first letter comes first in the alphabet. If the two Strings both begin with the 
same letter, then the method should return the shortest between the two. If they 
begin with the same letter and have the same length, then the method should 
return either one of them. And a method called reverseString that takes one String as input and 
returns a new String containing the same characters in reverse order.
 For example, if the input is “bananagrams”, the method should return 
“smargananab”